Refatora tabelas de pedidos e materiais

+++ database/migrations/2022_06_10_191314_create_materials_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->string('name');
             $table->integer('amount');
-            $table->double('price');
+            $table->float('price');
             $table->timestamps();
         });
     }
